--- src/About.tsx
@@ -1,10 +1,0 @@
-const About = () => {
-    return (
-        <div>
-            <h1>About Page</h1>
-            <p>This is the About page of your application.</p>
-        </div>
-    );
-};
-
-export default About;
