+++ src/components/Navbar.tsx
@@ -1,165 +1,28 @@
-"use client"
+import Header from "@/components/Header";
 
-import * as React from "react"
- 
-import { cn } from "@/lib/utils"
-import { Icons } from "@/components/icons"
-import {
-  NavigationMenu,
-  NavigationMenuContent,
-  NavigationMenuItem,
-  NavigationMenuLink,
-  NavigationMenuList,
-  NavigationMenuTrigger,
-  navigationMenuTriggerStyle,
-} from "@/components/ui/navigation-menu"
-
-const intro: { title: string; href: string; description: string }[] = [
-  {
-    title: "Alert Lịch sử phát triển",
-    href: "/docs/primitives/alert-dialog",
-    description:
-      "Lich sử phát triển của HealthCare",
-  },
-  {
-    title: "Đội ngũ nhân viên",
-    href: "/docs/primitives/alert-dialog",
-    description:
-      "Đội ngũ nhân viên của HealthCare",
-  },
-]
-
-const shops: { title: string; href: string; description: string }[] = [
-  {
-    title: "Dịch Vụ 1",
-    href: "",
-    description:
-      "Dịch Vụ 1 description",
-  },
-  {
-    title: "Dịch Vụ 2",
-    href: "",
-    description:
-      "Dịch Vụ 2 description",
-  },
-  {
-    title: "Dịch Vụ 3",
-    href: "",
-    description:
-      "Dịch Vụ 3 description",
-  },
-  
-]
-
-export default function Header() {
+import { Button } from "@/components/ui/button";
+import Paw from "@/assets/Paw2.svg";
+import { Link } from "react-router-dom";
+const Navbar = () => {
   return (
-    <div className="mt-[2.25rem] ml-10">
-      <NavigationMenu>
-        <NavigationMenuList>
-          <NavigationMenuItem>
-            <NavigationMenuTrigger className="bg-[#F3F4F6]">Getting started</NavigationMenuTrigger>
-            <NavigationMenuContent>
-              <ul className="grid gap-3 p-6 md:w-[400px] lg:w-[500px] lg:grid-cols-[.75fr_1fr]">
-                <li className="row-span-3">
-                  <NavigationMenuLink asChild>
-                    <a
-                      className="flex h-full w-full select-none flex-col justify-end rounded-md bg-gradient-to-b from-muted/50 to-muted p-6 no-underline outline-none focus:shadow-md"
-                      href="/"
-                    >
-                      <Icons.logo className="h-6 w-6" />
-                      <div className="mb-2 mt-4 text-lg font-medium">
-                        shadcn/ui
-                      </div>
-                      <p className="text-sm leading-tight text-muted-foreground">
-                        Beautifully designed components that you can copy and
-                        paste into your apps. Accessible. Customizable. Open
-                        Source.
-                      </p>
-                    </a>
-                  </NavigationMenuLink>
-                </li>
-                <ListItem href="/docs" title="Introduction">
-                  Re-usable components built using Radix UI and Tailwind CSS.
-                </ListItem>
-                <ListItem href="/docs/installation" title="Installation">
-                  How to install dependencies and structure your app.
-                </ListItem>
-                <ListItem href="/docs/primitives/typography" title="Typography">
-                  Styles for headings, paragraphs, lists...etc
-                </ListItem>
-              </ul>
-            </NavigationMenuContent>
-          </NavigationMenuItem>
-          <NavigationMenuItem>
-            <NavigationMenuTrigger className="bg-[#F3F4F6]">Giới Thiệu</NavigationMenuTrigger>
-            <NavigationMenuContent>
-              <ul className="grid w-[400px] gap-3 p-4 md:w-[500px] md:grid-cols-2 lg:w-[600px] ">
-                {intro.map((component) => (
-                  <ListItem
-                    key={component.title}
-                    title={component.title}
-                    href={component.href}
-                  >
-                    {component.description}
-                  </ListItem>
-                ))}
-              </ul>
-            </NavigationMenuContent>
-          </NavigationMenuItem>
-          <NavigationMenuItem>
-            <NavigationMenuTrigger className="bg-[#F3F4F6]">Dịch Vụ</NavigationMenuTrigger>
-            <NavigationMenuContent>
-              <ul className="grid w-[300px] gap-3 p-4 md:w-[400px] md:grid-cols-2 lg:w-[500px] ">
-                {shops.map((component) => (
-                  <ListItem
-                    key={component.title}
-                    title={component.title}
-                    href={component.href}
-                  >
-                    {component.description}
-                  </ListItem>
-                ))}
-              </ul>
-            </NavigationMenuContent>
-          </NavigationMenuItem>
-          <NavigationMenuItem>
-              <NavigationMenuLink className={`${navigationMenuTriggerStyle()} bg-[#F3F4F6]`}>
-                Thông Tin
-              </NavigationMenuLink>
-          </NavigationMenuItem>
-          <NavigationMenuItem>
-              <NavigationMenuLink className={`${navigationMenuTriggerStyle()} bg-[#F3F4F6]`}>
-                Liên Hệ
-              </NavigationMenuLink>
-          </NavigationMenuItem>
-        </NavigationMenuList>
-      </NavigationMenu>
+    <div>
+      <div className="flex bg-[#F3F4F6] pb-8 justify-between ">
+        <div className="flex">
+          <img
+            src={Paw}
+            className="ml-[3rem] mt-[2rem] w-[50px] h-[50px]"
+            alt=""
+          />
+          <Header />
+        </div>
+        <div className="mr-[3rem]">
+          <Button className="bg-[#DB2777] mt-[2.25rem] justify-self-end">
+            <Link to={"/login"}>Login</Link>
+          </Button>
+        </div>
+      </div>
     </div>
-  )
-}
- 
-const ListItem = React.forwardRef<
-  React.ElementRef<"a">,
-  React.ComponentPropsWithoutRef<"a">
->(({ className, title, children, ...props }, ref) => {
-  return (
-    <li>
-      <NavigationMenuLink asChild>
-        <a
-          ref={ref}
-          className={cn(
-            "block select-none space-y-1 rounded-md p-3 leading-none no-underline outline-none transition-colors hover:bg-accent hover:text-accent-foreground focus:bg-accent focus:text-accent-foreground",
-            className
-          )}
-          {...props}
-        >
-          <div className="text-sm font-medium leading-none">{title}</div>
-          <p className="line-clamp-2 text-sm leading-snug text-muted-foreground">
-            {children}
-          </p>
-        </a>
-      </NavigationMenuLink>
-    </li>
-  )
-})
-ListItem.displayName = "ListItem"
+  );
+};
+
+export default Navbar;
