+++ src/pages/About.tsx
@@ -1,0 +1,12 @@
+import Navbar from '../components/Header';
+
+const About = () => {
+    return (
+        <div>
+            <Navbar />
+            <h1>About page</h1>
+        </div>
+    );
+};
+
+export default About;
