--- src/pages/Dashboard.css
@@ -1,241 +1,0 @@
-/*
-color 
-#dd327e : pink
-#fce7f3 : light pink
-#fdf2f8 : light pink
-#f4f4f4 : light grey
-#f0e4f1 : light pink
-#eceff5 : light blue
-
-*/
-
-body {
-  font-family: "Arial", sans-serif;
-  margin: 0;
-  padding: 0;
-  background-color: #f4f4f4;
-  height: 100vh;
-  overflow: hidden;
-}
-
-.container {
-  display: flex;
-  width: 100vw;
-  height: 100vh;
-  border-radius: 10px;
-  background-color: #fce7f3;
-}
-
-.sidebar {
-  color: #dd327e;
-  width: 15vw;
-  height: 100%;
-  padding: 20px;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  background-color: #fdf2f8;
-  overflow-y: auto; /* Allow scrolling in the sidebar if needed */
-}
-
-.profile {
-  text-align: center;
-  margin-bottom: 20px;
-}
-
-.profile img {
-  display: block;
-  margin: 0 auto;
-  border-radius: 50%;
-  width: 80px;
-  height: 80px;
-}
-
-.profile h2 {
-  margin: 10px 0 5px;
-}
-
-.nav-menu {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-.nav-menu ul {
-  list-style: none;
-  padding: 0;
-  width: 100%;
-}
-
-.nav-menu li {
-  margin: 20px 0;
-}
-
-.nav-menu li:hover {
-  background-color: #f0e4f1;
-  border-radius: 5px;
-}
-
-.nav-menu a {
-  color: #dd327e;
-  text-decoration: none;
-  font-size: 18px;
-  font-weight: bold;
-  display: block;
-  width: 100%;
-  text-align: center;
-}
-
-/* Content */
-.content {
-  display: flex;
-  flex-direction: column; /* Ensure content stacks vertically on mobile */
-  width: 85vw;
-  height: 100%;
-  padding: 20px;
-  box-sizing: border-box;
-  overflow-y: auto; /* Allow scrolling in the content area */
-}
-
-.dashboard-boards {
-  display: flex;
-  justify-content: center;
-  margin-bottom: 20px;
-  gap: 20px;
-}
-
-.accounts-board,
-.revenue-board {
-  color: #dd327e;
-  background-color: #eceff5;
-  padding: 20px;
-  border-radius: 10px;
-  text-align: center;
-  flex: 1; /* Allow boards to take equal space */
-}
-
-.accounts-count,
-.revenue-amount {
-  color: #dd327e;
-  font-size: 24px;
-  font-weight: bold;
-  margin-top: 10px;
-}
-
-.main-content {
-  flex: 1; /* Allow main-content to take available space */
-  padding: 20px;
-  margin-bottom: 20px;
-  border-radius: 10px;
-  background-color: #fff;
-  overflow-y: auto; /* Allow scrolling within the main-content */
-}
-
-.sale-revenue {
-  flex: none; /* Prevent sale-revenue from taking additional space */
-  width: 100%;
-  background-color: #eceff5;
-  border-radius: 10px;
-}
-
-header {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  margin-bottom: 20px;
-}
-
-header h1 {
-  margin: 0;
-  font-size: 24px;
-}
-
-.header-icons img {
-  border-radius: 50%;
-  width: 40px;
-  height: 40px;
-  margin-left: 10px;
-}
-
-.expense-list h2 {
-  margin-top: 0;
-  font-size: 20px;
-}
-
-.expense-list ul {
-  list-style: none;
-  padding: 0;
-}
-
-.expense-list li {
-  display: flex;
-  align-items: center;
-  padding: 10px 0;
-  border-bottom: 1px solid #e0e0e0;
-  white-space: nowrap; /* Prevent text wrapping */
-  gap: 10px;
-}
-
-/* Category */
-.category {
-  display: inline-block;
-  padding: 5px 10px;
-  border-radius: 5px;
-  color: #fff;
-  margin-right: 10px;
-  min-width: 10rem;
-}
-
-.category.hospitalization {
-  background: #00bcd4;
-}
-
-.category.booking {
-  background: #9c27b0;
-}
-
-.category.others {
-  background: #ff9800;
-}
-
-.time {
-  flex: 1;
-}
-
-.note {
-  flex: 2;
-}
-
-.amount {
-  font-weight: bold;
-}
-
-/* Media Queries for Mobile */
-@media (max-width: 768px) {
-  .container {
-    flex-direction: column;
-    height: auto;
-  }
-
-  .sidebar {
-    display: none;
-  }
-
-  .content {
-    flex-direction: column;
-    width: 100vw;
-    height: auto; /* Adjust height for mobile */
-  }
-
-  .main-content {
-    width: 100vw;
-    margin: 0;
-    overflow-y: auto; /* Ensure main-content can scroll */
-  }
-
-  .sale-revenue {
-    width: 100vw;
-    margin-top: 20px;
-    overflow-y: auto; /* Ensure sale-revenue can scroll */
-  }
-}
