+++ src/pages/Dashboard.tsx
@@ -1,8 +1,18 @@
+/*
+color
+#dd327e : pink
+#fce7f3 : light pink
+#fdf2f8 : light pink
+#f4f4f4 : light grey
+#f0e4f1 : light pink
+#eceff5 : light blue
+*/
 
 import { Link, Outlet, useLocation } from 'react-router-dom';
 import logo from "../assets/react.svg";
 import Report, { ReportItem } from '../components/Report';
-import './Dashboard.css';
+import Navbar from '../components/Navbar';
+
 const Dashboard = () => {
   const sampleData: ReportItem[] = [
     {
@@ -29,53 +39,60 @@ const Dashboard = () => {
   const isAccountsPage = location.pathname === '/dashboard/accounts';
 
   return (
-    <div className='container'>
-      <aside className='sidebar'>
-        <div className="profile">
-          <img src={logo} alt="profile picture" />
-          <h2>Admin Account</h2>
-          <p>email address</p>
-        </div>
-        <nav className='nav-menu'>
-          <ul>
-            <li><Link to="/dashboard">Dashboard</Link></li>
-            <li><Link to="/dashboard/revenue">Revenue</Link></li>
-            <li><Link to="/dashboard/accounts">Accounts</Link></li>
-            <li><Link to="/dashboard/settings">Settings</Link></li>
-          </ul>
-        </nav>
-      </aside>
-      <div className="content">
-        {!isAccountsPage && (
-          <div className="dashboard-boards">
-            <div className="accounts-board">
-              <h2>Total User Accounts</h2>
-              <div className="accounts-count">
-                {/* Display the total number of user accounts here */}
-                <span>1,234</span>
-              </div>
-            </div>
-            <div className="revenue-board">
-              <h2>Total Revenue</h2>
-              <div className="revenue-amount">
-                {/* Display the total revenue amount here */}
-                <span>$123,456.78</span>
-              </div>
-            </div>
+    <div>
+      <div className="flex min-h-screen bg-gray-200">
+        <aside className="w-1/5 min-h-screen bg-pink-100 p-5 flex flex-col items-center overflow-y-auto">
+          <div className="text-center mb-5">
+            <img src={logo} alt="profile picture" className="mx-auto rounded-full w-20 h-20" />
+            <h2 className="text-pink-600 mt-2">Admin Account</h2>
+            <p>email address</p>
           </div>
-        )}
-        {!isAccountsPage && (
-          <div className='main-content'>
+          <nav className="w-full">
+            <ul className="space-y-5">
+              <li>
+                <Link to="/dashboard" className="text-pink-600 font-bold hover:bg-pink-200 rounded px-2 py-1 block">Dashboard</Link>
+              </li>
+              <li>
+                <Link to="/dashboard/revenue" className="text-pink-600 font-bold hover:bg-pink-200 rounded px-2 py-1 block">Revenue</Link>
+              </li>
+              <li>
+                <Link to="/dashboard/accounts" className="text-pink-600 font-bold hover:bg-pink-200 rounded px-2 py-1 block">Accounts</Link>
+              </li>
+              <li>
+                <Link to="/dashboard/settings" className="text-pink-600 font-bold hover:bg-pink-200 rounded px-2 py-1 block">Settings</Link>
+              </li>
+              {/* Link to home page */}
+              <li>
+                <Link to ="/" className="text-pink-600 font-bold hover:bg-pink-200 rounded px-2 py-1 block">Home Page</Link>
+              </li>
+            </ul>
+          </nav>
+        </aside>
+        <div className="w-4/5 p-5 flex flex-col overflow-y-auto">
+          {!isAccountsPage && (
+            <div className="flex justify-center mb-5 gap-5">
+              <div className="bg-blue-100 text-pink-600 p-5 rounded-lg flex-1">
+                <h2 className="text-xl font-bold mb-2">Total User Accounts</h2>
+                <div className="text-2xl font-bold">1,234</div>
+              </div>
+              <div className="bg-blue-100 text-pink-600 p-5 rounded-lg flex-1">
+                <h2 className="text-xl font-bold mb-2">Total Revenue</h2>
+                <div className="text-2xl font-bold">$123,456.78</div>
+              </div>
+            </div>
+          )}
+          {!isAccountsPage && (
+            <div className="bg-white p-5 rounded-lg mb-5 flex-1 overflow-y-auto">
             <div className="expense-menu">
-              <header>
+              <header className="flex justify-between items-center mb-5">
                 <div className='header-content'>
-                  <h1>Revenue</h1>
+                  <h1 className="text-2xl font-bold text-pink-600">Revenue</h1>
                 </div>
-                <p>20 - 5 - 2024</p>
+                <p className="text-gray-500">20 - 5 - 2024</p>
               </header>
               <section className="expense-details">
-                <div className="expense-chart">
-                  <div className="bar-chart">
+                <div className="expense-chart mb-5">
+                  <div className="bar-chart bg-gray-200 p-5 rounded-lg">
                     {/* Add bar chart implementation here */}
                   </div>
                 </div>
@@ -86,8 +103,9 @@ const Dashboard = () => {
               </section>
             </div>
           </div>
-        )}
-        {isAccountsPage && <Outlet />}
+          )}
+          {isAccountsPage && <Outlet />}
+        </div>
       </div>
     </div>
   );
